+++ CHANGELOG.md
@@ -6,11 +6,19 @@
 
 ### Changed
 
-- `extendedGradle` affects `testFunctional`, `testIntegration`, and `testFixtures` source sets 
-
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+### Updated
+
+## [0.4.0-alpha.5] - 2023-01-26
+
+### Changed
+
+- `extendedGradle` affects `testFunctional`, `testIntegration`, and `testFixtures` source sets
 
 ### Fixed
 
@@ -143,7 +151,9 @@
 - `org.eclipse.jgit:org.eclipse.jgit -> 6.4.0.202211300538-r`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.46`
 
-[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.4.0-alpha.4...HEAD
+[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.4.0-alpha.5...HEAD
+
+[0.4.0-alpha.5]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.4.0-alpha.4...0.4.0-alpha.5
 
 [0.4.0-alpha.4]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.4.0-alpha.3...0.4.0-alpha.4